Complete Override Manifest for react-native-win32
This change fills out the parts of the override manifest that tooling couldn't figure out. Quite a few overides no longer exist upstream, or are no longer needed, so we get to remove them here. There are a couple of other trivial changes to remove some overrides.

We make liberal use of the "derived" type here, even for simple stubs. This means that our stubs will follow Facebook stubs, and we will get a lint break (once validation is enabled) if Facebook removes files. This is seemingly valuable, given how many of our stubs don't track anything upstream.

Validated that we can bundle, build, linting passes, and that override validation now passes.
