+++ packages/react-native-win32/.flowconfig
@@ -18,7 +18,6 @@
 <PROJECT_ROOT>/Libraries/Components/StatusBar/StatusBar.js
 <PROJECT_ROOT>/Libraries/Components/TextInput/TextInput.js
 <PROJECT_ROOT>/Libraries/Components/TextInput/TextInputState.js
-<PROJECT_ROOT>/Libraries/Components/TimePickerAndroid/TimePickerAndroid.js
 <PROJECT_ROOT>/Libraries/Components/Touchable/TouchableNativeFeedback.js
 <PROJECT_ROOT>/Libraries/Components/View/ReactNativeViewAttributes.js
 <PROJECT_ROOT>/Libraries/Components/View/ReactNativeViewViewConfig.js
