+++ packages/react-native-win32/src/Libraries/Components/DrawerAndroid/DrawerLayoutAndroid.tsx
@@ -3,7 +3,8 @@
  */
 
 'use strict';
-import UnimplementedView from '../UnimplementedStubs/UnimplementedViews';
+// @ts-ignore This file is present after the JS build
+import * as UnimplementedView from '../UnimplementedViews/UnimplementedView';
 import { DrawerLayoutAndroidProps } from 'react-native';
 
 class DrawerLayoutAndroid extends UnimplementedView<DrawerLayoutAndroidProps> {}
