+++ packages/react-native-win32/src/Libraries/Components/Picker/PickerAndroid.tsx
@@ -3,7 +3,8 @@
  */
 
 'use strict';
-import UnimplementedView from '../UnimplementedStubs/UnimplementedViews';
+// @ts-ignore This file is present after the JS build
+import * as UnimplementedView from '../UnimplementedViews/UnimplementedView';
 import { PickerPropsAndroid } from 'react-native';
 
 class PickerAndroid extends UnimplementedView<PickerPropsAndroid> {}
