+++ packages/react-native-win32/src/Libraries/Components/StatusBar/StatusBar.win32.ts
@@ -1,5 +1,10 @@
 /**
- * This copies what StatusBar.android.js does
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @format
  */
 'use strict';
