--- packages/react-native-win32/src/Libraries/Components/TimePickerAndroid/TimePickerAndroid.win32.js
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- * @format
- */
-'use strict';
-
-module.exports = require('../UnimplementedViews/UnimplementedView');
