--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/ActivityIndicatorIOS.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule ActivityIndicatorIOS
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { ActivityIndicatorIOSProps } from 'react-native';
-
-class ActivityIndicatorIOS extends UnimplementedView<ActivityIndicatorIOSProps> {}
-
-export = ActivityIndicatorIOS;
