--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/IntentAndroid.win32.ts
@@ -1,10 +1,0 @@
-/**
- * @providesModule IntentAndroid
- */
-'use strict';
-
-export = {
-  openURL: url => {
-    console.error('IntentAndroid is not supported on iOS');
-  }
-};
