--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/NavigationHeaderBackButton.win32.ts
@@ -1,7 +1,0 @@
-/**
- * @providesModule NavigationHeaderBackButton
- * @flow
- */
-'use strict';
-
-export = null;
