--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/NavigatorBreadcrumbNavigationBarStyles.win32.ts
@@ -1,19 +1,0 @@
-/**
- * @providesModule NavigatorBreadcrumbNavigationBarStyles
- */
-'use strict';
-
-/**
- * Contains constants that are used in constructing both `StyleSheet`s and
- * inline styles during transitions.
- */
-export = {
-  Interpolators: null,
-  Left: null,
-  Center: null,
-  Right: null,
-  IconWidth: null,
-  IconHeight: null,
-  SeparatorWidth: null,
-  SeparatorHeight: null
-};
