--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/NavigatorIOS.tsx
@@ -1,12 +1,0 @@
-/**
- * @providesModule NavigatorIOS
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { ViewProps } from 'react-native';
-
-// Should be NavigatorIOSProperties?
-class NavigatorIOS extends UnimplementedView<ViewProps> {}
-
-export = NavigatorIOS;
