--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/SliderIOS.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule SliderIOS
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { SliderPropsIOS } from 'react-native';
-
-class SliderIOS extends UnimplementedView<SliderPropsIOS> {}
-
-export = SliderIOS;
