--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/SnapshotViewIOS.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule SnapshotViewIOS
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { SnapshotViewIOSProps } from 'react-native';
-
-class SnapshotViewIOS extends UnimplementedView<SnapshotViewIOSProps> {}
-
-export = SnapshotViewIOS;
