--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/SwitchAndroid.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule SwitchAndroid
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { SwitchProps } from 'react-native';
-
-class SwitchAndroid extends UnimplementedView<SwitchProps> {}
-
-export = SwitchAndroid;
