--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/SwitchIOS.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule SwitchIOS
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { SwitchIOSProps } from 'react-native';
-
-class SwitchIOS extends UnimplementedView<SwitchIOSProps> {}
-
-export = SwitchIOS;
