--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/TabBarIOS.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule TabBarIOS
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { TabBarIOSProps } from 'react-native';
-
-class TabBarIOS extends UnimplementedView<TabBarIOSProps> {}
-
-export = TabBarIOS;
