--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/ToolbarAndroid.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule ToolbarAndroid
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { ToolbarAndroidProps } from 'react-native';
-
-class ToolbarAndroid extends UnimplementedView<ToolbarAndroidProps> {}
-
-export = ToolbarAndroid;
