--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/VibrationIOS.win32.ts
@@ -1,12 +1,0 @@
-/**
- * @providesModule VibrationIOS
- */
-'use strict';
-
-// var warning = require('fbjs/lib/warning');
-
-export = {
-  vibrate: () => {
-    // warning('VibrationIOS is not supported on this platform!');
-  }
-};
