--- packages/react-native-win32/src/Libraries/Components/UnimplementedStubs/ViewPagerAndroid.tsx
@@ -1,11 +1,0 @@
-/**
- * @providesModule ViewPagerAndroid
- */
-
-'use strict';
-import UnimplementedView from './UnimplementedViews';
-import { ViewPagerAndroidProps } from 'react-native';
-
-class ViewPagerAndroid extends UnimplementedView<ViewPagerAndroidProps> {}
-
-export = ViewPagerAndroid;
