--- packages/react-native-win32/src/Libraries/Utilities/BackAndroid.win32.ts
@@ -1,28 +1,0 @@
-/**
- * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- *
- * iOS stub for BackAndroid.android.js
- *
- * @providesModule BackAndroid
- */
-
-'use strict';
-
-/* tslint:disable:no-empty */
-function emptyFunction() {}
-/* tslint:enable:no-empty */
-
-export = {
-  exitApp: emptyFunction,
-  addEventListener() {
-    return {
-      remove: emptyFunction
-    };
-  },
-  removeEventListener: emptyFunction
-};
